requirements.txt
absl-py==0.5.0
astor==0.7.1
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
Keras==2.1.6
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
Markdown==3.0.1
numpy==1.14.1
protobuf==3.6.1
PyYAML==3.13
scipy==1.1.0
six==1.11.0
tensorboard==1.9.0
tensorflow==1.9.0
tensorflow-hub==0.1.0
tensorflowjs==0.5.7
termcolor==1.1.0
Werkzeug==0.14.1
Pillow==5.3.0
joblib==0.12.5
